requirements.txt
langchain
langchain_community
langchainhuggingface_hub
streamlit
faiss-cpu
pypdf